Admin Middleware added

+++ routes/web.php
@@ -75,13 +75,11 @@ Route::post('/user/{username}', 'UserController@update')->name('user-update');
 
 
 // Admin/Moderators
-Route::get('/admin/{category}', 'AdminController@show')->name('adminPage');
-
-
-Route::post('/admin/block/{user}', 'AdminController@changeBlock')->name('changeBlock');
-Route::delete('/admin/users/{user}', 'AdminController@deleteUser')->name('deleteUser');
-Route::delete('/admin/questions/{question}', 'AdminController@deleteQuestion')->name('deleteQuestion');
-Route::delete('/admin/tags/{tag}', 'AdminController@deleteTag')->name('deleteTag');
+Route::get('/admin/{category}', 'AdminController@show')->name('adminPage')->middleware('admin');
+Route::post('/admin/block/{user}', 'AdminController@changeBlock')->name('changeBlock')->middleware('admin');
+Route::delete('/admin/users/{user}', 'AdminController@deleteUser')->name('deleteUser')->middleware('admin');
+Route::delete('/admin/questions/{question}', 'AdminController@deleteQuestion')->name('deleteQuestion')->middleware('admin');
+Route::delete('/admin/tags/{tag}', 'AdminController@deleteTag')->name('deleteTag')->middleware('admin');
